+++ libs/velo-external-db-core/src/utils/schema_utils.spec.ts
@@ -61,7 +61,7 @@ describe('Schema utils functions', () => {
             expect(subtypeToFieldType(VeloFieldTypeEnum.text)).toBe('string')
         })
         test('number type', () => {
-            expect(subtypeToFieldType(VeloFieldTypeEnum.number)).toBe('int')
+            expect(subtypeToFieldType(VeloFieldTypeEnum.number)).toBe('float')
         })
         test('boolean type', () => {
             expect(subtypeToFieldType(VeloFieldTypeEnum.boolean)).toBe('')
